requirements.txt
# LinkedIn Network Intelligence - Python Dependencies
# Install: pip install -r requirements.txt

# =============================================================================
# Core Dependencies
# =============================================================================
pydantic>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
pandas>=2.0.0

# =============================================================================
# LLM Providers
# =============================================================================
anthropic>=0.40.0
openai>=1.0.0
httpx>=0.25.0  # For Ollama/vLLM HTTP calls

# =============================================================================
# CLI & User Interface
# =============================================================================
click>=8.0.0
rich>=13.0.0  # Pretty console output and progress bars

# =============================================================================
# Caching & Storage
# =============================================================================
diskcache>=5.6.0  # SQLite-backed caching

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0  # .env file support

# =============================================================================
# Development Dependencies (install with: pip install -r requirements.txt[dev])
# =============================================================================
# Uncomment for development:
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
# mypy>=1.0.0
# ruff>=0.1.0
# black>=23.0.0

# =============================================================================
# Optional: Local LLM Support
# =============================================================================
# For vLLM (requires CUDA):
# vllm>=0.2.0

# For direct transformers usage:
# transformers>=4.35.0
# torch>=2.0.0
